requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
antiorm==1.2.1
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.11.1
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
creart==0.2.2
creart-graia==0.1.5 ; python_version >= '3.8'
croniter==1.3.8 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
db==0.1.1
db-sqlite3==0.0.1
frozenlist==1.3.3 ; python_version >= '3.7'
graia-amnesia==0.7.1 ; python_version >= '3.8' and python_version < '4.0'
graia-ariadne[graia]==0.10.3
graia-broadcast==0.19.1 ; python_version >= '3.7' and python_version < '4.0'
graia-saya==0.0.17
graia-scheduler==0.1.4
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version >= '3.7'
launart==0.6.3 ; python_version >= '3.8'
loguru==0.6.0 ; python_version >= '3.5'
lxml==4.9.2
multidict==6.0.4 ; python_version >= '3.7'
pydantic==1.10.4 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0
requests==2.28.2
ruamel.yaml==0.17.21
schedule==1.1.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
statv==0.3.2 ; python_version >= '3.8'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.8.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
